test: util method
