+++ src/components/Markdown/Markdown.tsx
@@ -21,16 +21,6 @@ interface MarkdownProps {
   variant?: "plain" | "subtitle" | "legal";
 }
 
-const getLinkAction = (href: string) => {
-  if (href === "#button-cookie-edit-preferences") {
-    return () => {
-      document.getElementById("ot-sdk-btn-floating")?.click();
-    };
-  }
-
-  return undefined;
-};
-
 const subtitleComponents: Components = {
   p({ children }) {
     return <FormControlDescription>{children}</FormControlDescription>;
